+++ common/src/main/java/com/railwayteam/railways/mixin/MixinScheduleItem.java
@@ -22,7 +22,7 @@ public abstract class MixinScheduleItem {
         if (pInteractionTarget.getRootVehicle() instanceof CarriageContraptionEntity cce) {
             Carriage carriage = cce.getCarriage();
             Train train = carriage.train;
-            ((IIndexedSchedule) train).setIndex(train.carriages.indexOf(carriage));
+            ((IIndexedSchedule) train).snr$setIndex(train.carriages.indexOf(carriage));
         }
     }
 }