+++ changelog.md
@@ -5,10 +5,17 @@ Additions
 - Cherry Tracks (1.20.1)
 - Bamboo Tracks (1.20.1)
 - Stripped Bamboo Tracks (1.20.1)
+- Handcar: wrench to pick up
 - Players can now click a station with a whistle to summon a train directly to that station
 - A deployer using a whistle on air will clear the schedule of the bound train
 - You can now whistle a dual-headed train while in motion and have it slow and reverse
+- Config option to use old-style smoke (campfire particles) for smokestacks
+
+Changes
+- Generic Crossing Tracks - any two tracks can now cross at a junction
 - Decoupled trains' names will now stop at one "Split off from: ..."
+- Trains will now approach nearby stations when they are the rear train in a decoupled set
+- Trains relocated by a wrench will now try to approach nearby stations after relocation
 - Fix Farmers delight fabric crash (small hacky fix)
 ------------------------------------------------------
 Version 1.5.1